Refactored inspector logic into secondary manager classes
