+++ src/main/groovy/com/blackducksoftware/integration/hub/detect/bomtool/DockerBomTool.groovy
@@ -23,7 +23,6 @@
 package com.blackducksoftware.integration.hub.detect.bomtool
 
 import org.apache.commons.io.IOUtils
-import org.apache.commons.lang3.math.NumberUtils
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
@@ -36,27 +35,21 @@ import com.blackducksoftware.integration.hub.bdio.model.Forge
 import com.blackducksoftware.integration.hub.bdio.model.SimpleBdioDocument
 import com.blackducksoftware.integration.hub.bdio.model.externalid.ExternalId
 import com.blackducksoftware.integration.hub.bdio.model.externalid.ExternalIdFactory
+import com.blackducksoftware.integration.hub.detect.bomtool.docker.DockerInspectorManager
 import com.blackducksoftware.integration.hub.detect.bomtool.docker.DockerProperties
 import com.blackducksoftware.integration.hub.detect.hub.HubSignatureScanner
 import com.blackducksoftware.integration.hub.detect.model.BomToolType
 import com.blackducksoftware.integration.hub.detect.model.DetectCodeLocation
 import com.blackducksoftware.integration.hub.detect.type.ExecutableType
 import com.blackducksoftware.integration.hub.detect.util.executable.Executable
-import com.blackducksoftware.integration.hub.rest.UnauthenticatedRestConnection
-import com.blackducksoftware.integration.log.Slf4jIntLogger
 import com.google.gson.Gson
 
 import groovy.transform.TypeChecked
-import okhttp3.HttpUrl
-import okhttp3.Request
-import okhttp3.Response
 
 @Component
 @TypeChecked
 class DockerBomTool extends BomTool {
     private final Logger logger = LoggerFactory.getLogger(DockerBomTool.class)
-
-    static final URL LATEST_URL = new URL('https://blackducksoftware.github.io/hub-docker-inspector/hub-docker-inspector.sh')
 
     final String tarFilenamePattern = '*.tar.gz'
     final String dependenciesFilenamePattern = '*bdio.jsonld'
@@ -76,10 +69,11 @@ class DockerBomTool extends BomTool {
     @Autowired
     BdioTransformer bdioTransformer
 
+    @Autowired
+    DockerInspectorManager dockerInspectorManager
+
     private String dockerExecutablePath
     private String bashExecutablePath
-    private File dockerInspectorShellScript
-    private String inspectorVersion
 
     @Override
     public BomToolType getBomToolType() {
@@ -111,9 +105,7 @@ class DockerBomTool extends BomTool {
         File dockerBomToolDirectory =  dockerPropertiesFile.getParentFile()
         dockerProperties.populatePropertiesFile(dockerPropertiesFile, dockerBomToolDirectory)
 
-        if(!dockerInspectorShellScript) {
-            dockerInspectorShellScript = getShellScript()
-        }
+        File dockerInspectorShellScript = dockerInspectorManager.getShellScript()
 
         boolean usingTarFile = false
         String imageArgument = ''
@@ -134,7 +126,7 @@ class DockerBomTool extends BomTool {
             '-c',
             "\"${dockerInspectorShellScript.getCanonicalPath()}\" --spring.config.location=\"${dockerBomToolDirectory.getCanonicalPath()}\" --dry.run=true --no.prompt=true ${imageArgument}" as String
         ]
-        def airGapHubDockerInspectorJar = new File("${detectConfiguration.getDockerInspectorAirGapPath()}", "hub-docker-inspector-${inspectorVersion}.jar")
+        def airGapHubDockerInspectorJar = new File("${detectConfiguration.getDockerInspectorAirGapPath()}", "hub-docker-inspector-${dockerInspectorManager.getInspectorVersion(bashExecutablePath)}.jar")
         if(airGapHubDockerInspectorJar.exists()) {
             try {
                 for ( String os : ['ubuntu', 'alpine', 'centos']) {
@@ -197,65 +189,7 @@ class DockerBomTool extends BomTool {
     }
 
     String getInspectorVersion() {
-        if ('latest'.equalsIgnoreCase(detectConfiguration.getDockerInspectorVersion())) {
-            if (!inspectorVersion) {
-                File dockerPropertiesFile = detectFileManager.createFile(getBomToolType(), 'application.properties')
-                File dockerBomToolDirectory =  dockerPropertiesFile.getParentFile()
-                if(!dockerInspectorShellScript) {
-                    dockerInspectorShellScript = getShellScript()
-                }
-                List<String> bashArguments = [
-                    '-c',
-                    "\"${dockerInspectorShellScript.getCanonicalPath()}\" --version" as String
-                ]
-                Executable getDockerInspectorVersion = new Executable(dockerBomToolDirectory, bashExecutablePath, bashArguments)
-
-                inspectorVersion = executableRunner.execute(getDockerInspectorVersion).standardOutput.split(' ')[1]
-            }
-        } else {
-            inspectorVersion = detectConfiguration.getDockerInspectorVersion()
-        }
-        inspectorVersion
-    }
-
-    private File getShellScript() {
-        File shellScriptFile
-        def airGapHubDockerInspectorShellScript = new File(detectConfiguration.getDockerInspectorAirGapPath(), 'hub-docker-inspector.sh')
-        logger.debug("Verifying air gap shell script present at ${airGapHubDockerInspectorShellScript.getCanonicalPath()}")
-
-        if (detectConfiguration.dockerInspectorPath) {
-            shellScriptFile = new File(detectConfiguration.dockerInspectorPath)
-        } else if (airGapHubDockerInspectorShellScript.exists()) {
-            shellScriptFile = airGapHubDockerInspectorShellScript
-        } else {
-            URL hubDockerInspectorShellScriptUrl = LATEST_URL
-            if (!'latest'.equals(detectConfiguration.dockerInspectorVersion)) {
-                hubDockerInspectorShellScriptUrl = new URL("https://blackducksoftware.github.io/hub-docker-inspector/hub-docker-inspector-${detectConfiguration.dockerInspectorVersion}.sh")
-            }
-            logger.info("Getting the Docker inspector shell script from ${hubDockerInspectorShellScriptUrl.toURI().toString()}")
-            UnauthenticatedRestConnection restConnection = new UnauthenticatedRestConnection(new Slf4jIntLogger(logger), hubDockerInspectorShellScriptUrl, detectConfiguration.getHubTimeout())
-            restConnection.alwaysTrustServerCertificate = detectConfiguration.hubTrustCertificate
-            restConnection.proxyHost = detectConfiguration.getHubProxyHost()
-            restConnection.proxyPort = NumberUtils.toInt(detectConfiguration.getHubProxyPort())
-            restConnection.proxyUsername = detectConfiguration.getHubProxyUsername()
-            restConnection.proxyPassword = detectConfiguration.getHubProxyPassword()
-            HttpUrl httpUrl = restConnection.createHttpUrl()
-            Request request = restConnection.createGetRequest(httpUrl)
-            String shellScriptContents = null
-            Response response = null
-            try {
-                response = restConnection.handleExecuteClientCall(request)
-                shellScriptContents =  response.body().string()
-            } finally {
-                if (response != null) {
-                    response.close()
-                }
-            }
-            shellScriptFile = detectFileManager.createFile(getBomToolType(), "hub-docker-inspector-${detectConfiguration.dockerInspectorVersion}.sh")
-            detectFileManager.writeToFile(shellScriptFile, shellScriptContents)
-            shellScriptFile.setExecutable(true)
-        }
-        return shellScriptFile
+        return dockerInspectorManager.getInspectorVersion(bashExecutablePath)
     }
 
     private void logMissingFile(File searchDirectory, String filenamePattern) {
